+++ comicboxd/server/server.go
@@ -9,7 +9,7 @@ import (
 	"github.com/spf13/viper"
 	"github.com/zwzn/comicbox/comicboxd/app/database"
 	"github.com/zwzn/comicbox/comicboxd/j"
-	"github.com/zwzn/comicbox/comicboxd/tls"
+	"github.com/zwzn/comicbox/comicboxd/server/tls"
 
 	_ "github.com/golang-migrate/migrate/source/file"
 	"github.com/gorilla/mux"
